general update and adding of class, session, term, year

+++ app/Http/Controllers/AdminController.php
@@ -9,6 +9,9 @@ use App\Models\Module;
 use App\Models\ModuleRole;
 use App\Models\UserRole;
 use App\Models\ExamType;
+use App\Models\AcademicYear;
+use App\Models\AcademicSession;
+use App\Models\Semester;
 use App\Repositories\QuestionRepositoryInterface;
 use App\Repositories\AnswerRepositoryInterface;
 use App\Repositories\ModuleRoleRepositoryInterface;
@@ -16,6 +19,10 @@ use App\Repositories\AssignModuleRepositoryInterface;
 use App\Repositories\TakeExamRepositoryInterface;
 use App\Repositories\UserRepositoryInterface;
 use App\Repositories\ExamtypeRepositoryInterface;
+use App\Repositories\AcademicSessionInterface;
+use App\Repositories\AcademicYearInterface;
+use App\Repositories\SemesterInterface;
+use App\Repositories\ClassInterface;
 use DB;
 
 class AdminController extends Controller
@@ -28,8 +35,12 @@ class AdminController extends Controller
     private $takeexamRepository;
     private $userRepository;
     private $examtypeRepository;
+    private $semesterRepository;
+    private $sessionRepository;
+    private $yearRepository;
+    private $classRepository;
 
-    public function __construct(ExamtypeRepositoryInterface $examtypeRepository, UserRepositoryInterface $userRepository, TakeExamRepositoryInterface $takeexamRepository,AssignModuleRepositoryInterface $assignmoduleRepository, ModuleRoleRepositoryInterface $moduleroleRepository, QuestionRepositoryInterface $questionRepository,AnswerRepositoryInterface $answerRepository)
+    public function __construct(ClassInterface $classRepository, SemesterInterface $semesterRepository, AcademicSessionInterface $sessionRepository, AcademicYearInterface $yearRepository, ExamtypeRepositoryInterface $examtypeRepository, UserRepositoryInterface $userRepository, TakeExamRepositoryInterface $takeexamRepository,AssignModuleRepositoryInterface $assignmoduleRepository, ModuleRoleRepositoryInterface $moduleroleRepository, QuestionRepositoryInterface $questionRepository,AnswerRepositoryInterface $answerRepository)
     {
         $this->middleware('auth');
         $this->answerRepository = $answerRepository;
@@ -39,12 +50,27 @@ class AdminController extends Controller
         $this->takeexamRepository = $takeexamRepository;
         $this->userRepository = $userRepository;
         $this->examtypeRepository = $examtypeRepository;
+        $this->semesterRepository = $semesterRepository;
+        $this->sessionRepository = $sessionRepository;
+        $this->yearRepository = $yearRepository;
+        $this->classRepository = $classRepository;
     }
 
     public function viewQuestions()
     {
+        $data['sessionx']='';
+        $data['termx']='';
+        $data['yearx']='';
+        $data['examtypex']='';
+        $data['classx']='';
+
         $data['questions'] = $this->questionRepository->all();
+
         $data['examtype'] = $this->examtypeRepository->all();
+        $data['session'] = $this->sessionRepository->all();
+        $data['term'] = $this->semesterRepository->all();
+        $data['year'] = $this->yearRepository->all();
+        $data['class'] = $this->classRepository->all();
 
         return view('Questions.addQuestions',$data);
     }
@@ -53,20 +79,48 @@ class AdminController extends Controller
     public function saveQuestions(Request $request)
     {
         $this->validate( $request, [
-            'examtype' => 'required',
+            'class' => 'required',
+            'session' => 'required',
+            'term' => 'required|string',
+            'year' => 'required',
             'question' => 'required|string',
-        ]);
+            'examtype' => 'required',
+           ]);
         
-        $data['questions'] = $this->questionRepository->create(['examtype'=>$request->examtype,'question'=>$request->question,'score'=>$request->score]);
 
-        return back()->with('success','Question addedd successfully!');
+        $data['sessionx'] = $request->session;
+        $data['termx']  = $request->term;
+        $data['yearx']  = $request->year;
+        $data['examtypex']  = $request->examtype;
+        $data['classx'] =  $request->class;
+
+        
+        $data['examtype'] = $this->examtypeRepository->all();
+        $data['session'] = $this->sessionRepository->all();
+        $data['term'] = $this->semesterRepository->all();
+        $data['year'] = $this->yearRepository->all();
+        $data['class'] = $this->classRepository->all();
+        
+        if(DB::table('questions')->where('class',$request->class)->where('session',$request->session)->where('term',$request->term)->where('year',$request->year)
+        ->where('examtype',$request->examtype)->where('question',$request->question)->exists()){
+            return back()->with('error_message','Question already exists');
+        }else {
+        $data['questions'] = $this->questionRepository->create(['class'=>$request->class,'session'=>$request->session,'term'=>$request->term,'year'=>$request->year,'examtype'=>$request->examtype,'question'=>$request->question,'score'=>$request->score]);
+        }
+        $data['questions'] = $this->questionRepository->all();
+        //return back()->with('success','Question addedd successfully!');
+        return view('Questions.addQuestions',$data);
     }
 
     public function deleteQuestions($id)
     {   
-        $data['questions'] = $this->questionRepository->deleteQuestion(base64_decode($id));
-
-        return back()->with('success','Question deleted successfully!');
+        //dd($id);
+        if(DB::table('answers')->where('question_id',base64_decode($id))->exists()){
+            return back()->with('error_message','Questions already in used');
+        }else{
+            $data['questions'] = $this->questionRepository->deleteQuestion(base64_decode($id));
+        }
+            return back()->with('success','Question deleted successfully!');
     }
     
     //answers 
@@ -84,8 +138,13 @@ class AdminController extends Controller
     {
         $data['questionx'] = $request->question;
         $data['question'] = $this->questionRepository->all();
-        
+
+        if(DB::table('answers')->where('answer',$request->answer)->where('question_id',$request->question)->exists()){
+            return back()->with('error_message','Record already exists');
+        }else {
         $data['questions'] = $this->answerRepository->create(['question_id'=>$request->question,'answer'=>$request->answer,'correct_answer'=>$request->correct_answer]);
+        }
+
         $data['answer'] = $this->answerRepository->all();
 
         return view('Answers.addAnswers',$data);
@@ -93,8 +152,11 @@ class AdminController extends Controller
 
     public function deleteAnswers($id)
     {   
+        if(DB::table('take_exams')->where('answerID',base64_decode($id))->exists()){
+            return back()->with('error_message','Questions already in used');
+        }else{
         $data['questions'] = $this->answerRepository->deleteAnswer(base64_decode($id));
-
+        }
         return back()->with('success','Answer deleted successfully!');
     }
